UF-15168: Refactor db modell and add integration tests

+++ src/main/java/se/sundsvall/relations/service/RelationTypeService.java
@@ -7,40 +7,42 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.zalando.problem.Problem;
 import se.sundsvall.relations.api.model.RelationType;
+import se.sundsvall.relations.integration.db.RelationRepository;
 import se.sundsvall.relations.integration.db.RelationTypeRepository;
 import se.sundsvall.relations.service.mapper.RelationTypeMapper;
 
 @Service
 public class RelationTypeService {
 	private static final String NOT_FOUND_MSG = "Relation type with type '%s' not found";
-	private static final String DUPLICATE_TYPE_MSG = "Relation type with type '%s' already exists";
-	private static final String DUPLICATE_COUNTER_TYPE_MSG = "Relation type with counter type '%s' already exists";
 	private static final String SAME_TYPE_AND_COUNTER_TYPE_MSG = "Type and counter type cannot be the same: '%s'";
 	private static final String VALUE_ALREADY_EXISTS_MSG = "Value '%s' already exists as a type or counter type.";
+	private static final String TYPE_IS_USED_BY_RELATIONS_MSG = "Type '%s' is used by one or many Relations";
 
 	private final RelationTypeRepository relationTypeRepository;
+	private final RelationRepository relationRepository;
 	private final RelationTypeMapper mapper;
 
-	public RelationTypeService(RelationTypeRepository relationTypeRepository, RelationTypeMapper mapper) {
+	public RelationTypeService(RelationTypeRepository relationTypeRepository, RelationRepository relationRepository, RelationTypeMapper mapper) {
 		this.relationTypeRepository = relationTypeRepository;
+		this.relationRepository = relationRepository;
 		this.mapper = mapper;
 	}
 
 	public String createType(RelationType type) {
-		if (type.getType().equalsIgnoreCase(type.getCounterType())) {
-			throw Problem.valueOf(CONFLICT, SAME_TYPE_AND_COUNTER_TYPE_MSG.formatted(type.getType()));
+		if (type.getName().equalsIgnoreCase(type.getCounterName())) {
+			throw Problem.valueOf(CONFLICT, SAME_TYPE_AND_COUNTER_TYPE_MSG.formatted(type.getName()));
 		}
-		if (relationTypeRepository.existsByTypeOrCounterType(type.getType())) {
-			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(type.getType()));
+		if (relationTypeRepository.existsByName(type.getName())) {
+			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(type.getName()));
 		}
-		if (type.getCounterType() != null && relationTypeRepository.existsByTypeOrCounterType(type.getCounterType())) {
-			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(type.getCounterType()));
+		if (type.getCounterName() != null && relationTypeRepository.existsByName(type.getCounterName())) {
+			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(type.getCounterName()));
 		}
-		return relationTypeRepository.save(mapper.toRelationTypeEntity(type)).getId();
+		return relationTypeRepository.save(mapper.toRelationTypeEntity(type)).getName();
 	}
 
 	public RelationType getType(String type) {
-		return relationTypeRepository.findByType(type)
+		return relationTypeRepository.findByName(type)
 			.map(mapper::toRelationType)
 			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, NOT_FOUND_MSG.formatted(type)));
 	}
@@ -52,28 +54,43 @@ public class RelationTypeService {
 	}
 
 	public void deleteRelationType(String type) {
-		final var entity = relationTypeRepository.findByType(type)
+		final var entity = relationTypeRepository.findByName(type)
 			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, NOT_FOUND_MSG.formatted(type)));
+		if (relationRepository.existsByType(entity)) {
+			throw Problem.valueOf(CONFLICT, TYPE_IS_USED_BY_RELATIONS_MSG.formatted(entity.getName()));
+		}
 		relationTypeRepository.delete(entity);
 	}
 
 	public RelationType saveRelationType(String type, RelationType relationType) {
-		final var existingEntity = relationTypeRepository.findByType(type)
+		final var existingEntity = relationTypeRepository.findByName(type)
 			.orElseThrow(() -> Problem.valueOf(NOT_FOUND, NOT_FOUND_MSG.formatted(type)));
 
-		if (relationType.getType().equalsIgnoreCase(relationType.getCounterType())) {
-			throw Problem.valueOf(CONFLICT, SAME_TYPE_AND_COUNTER_TYPE_MSG.formatted(relationType.getType()));
+		if (relationType.getName().equalsIgnoreCase(relationType.getCounterName())) {
+			throw Problem.valueOf(CONFLICT, SAME_TYPE_AND_COUNTER_TYPE_MSG.formatted(relationType.getName()));
 		}
-		if (!existingEntity.getType().equalsIgnoreCase(relationType.getType()) && relationTypeRepository.existsByTypeOrCounterType(relationType.getType())) {
-			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(relationType.getType()));
+		if (!existingEntity.getName().equalsIgnoreCase(relationType.getName()) && relationTypeRepository.existsByName(relationType.getName())) {
+			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(relationType.getName()));
 		}
-		if (relationType.getCounterType() != null && !existingEntity.getCounterType().equalsIgnoreCase(relationType.getCounterType()) && relationTypeRepository.existsByTypeOrCounterType(relationType.getCounterType())) {
-			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(relationType.getCounterType()));
+
+		boolean relationCounterTypeExists = relationType.getCounterName() != null;
+		boolean existingCounterTypeIsDifferent = !(existingEntity.getCounterType() != null && existingEntity.getCounterType().getName().equalsIgnoreCase(relationType.getCounterName()));
+
+		if (relationCounterTypeExists && existingCounterTypeIsDifferent && relationTypeRepository.existsByName(relationType.getCounterName())) {
+			throw Problem.valueOf(CONFLICT, VALUE_ALREADY_EXISTS_MSG.formatted(relationType.getCounterName()));
 		}
 
 		final var updatedEntity = mapper.toRelationTypeEntity(relationType);
+
 		updatedEntity.setId(existingEntity.getId());
-		final var savedEntity = relationTypeRepository.save(updatedEntity);
-		return mapper.toRelationType(savedEntity);
+		if (existingEntity.getCounterType() != null) {
+			if (updatedEntity.getCounterType() != null) {
+				updatedEntity.getCounterType().setId(existingEntity.getCounterType().getId());
+			} else if (relationRepository.existsByType(existingEntity.getCounterType())) {
+				throw Problem.valueOf(CONFLICT, TYPE_IS_USED_BY_RELATIONS_MSG.formatted(existingEntity.getCounterType().getName()));
+			}
+		}
+
+		return mapper.toRelationType(relationTypeRepository.save(updatedEntity));
 	}
 }
